requirements.txt
absl-py==2.4.0
asttokens==3.0.1
cffi==2.0.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
flatbuffers==25.12.19
fonttools==4.61.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
mediapipe==0.10.32
MouseInfo==0.1.3
nest-asyncio==1.6.0
numpy==2.4.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==3.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pyzmq==27.1.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sounddevice==0.5.5
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
wcwidth==0.6.0